requirements.txt
Django==5.2.11
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.1
django-cors-headers==4.6.0
Pillow==10.4.0
python-decouple==3.8
psycopg2-binary==2.9.10
gunicorn==23.0.0
whitenoise==6.9.0
dj-database-url==2.3.0
cloudinary==1.41.0
django-cloudinary-storage==0.3.0
